Final CRUD

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -8,16 +8,18 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->uuid('id')->primary();
-            $table->string('name', 255);
-            $table->string('email', 255)->unique();
-            $table->string('password', 255);
-            $table->timestamp('email_verified_at')->nullable();
-            $table->uuid('role_id');
-            $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('users')) {
+            Schema::create('users', function (Blueprint $table) {
+                $table->uuid('id')->primary();
+                $table->string('name', 255);
+                $table->string('email', 255)->unique();
+                $table->string('password', 255);
+                $table->timestamp('email_verified_at')->nullable();
+                $table->uuid('role_id');
+                $table->foreign('role_id')->references('id')->on('roles')->onDelete('cascade');
+                $table->timestamps();
+            });
+        }
     }
 
     public function down(): void
